backend/requirements.txt
fastapi
uvicorn
python-dotenv
openai
python-multipart
pydantic
supabase>=2.25.0
websockets>=13.0
httpx==0.27.2
pypdf  